+++ app/components/modal-dialog.js
@@ -1,1 +1,1 @@
-export { default } from '@zestia/ember-modal-dialog/components/modal-dialog';
+export { default } from '@zestia/ember-modal-dialog/components/modal-dialog/index';