+++ app/templates/components/modal-dialog/content.js
@@ -1,1 +1,1 @@
-export { default } from '@zestia/ember-modal-dialog/components/modal-dialog/content/template';
+export { default } from '@zestia/ember-modal-dialog/components/modal-dialog/content';